Refactor: Extract _dispatchGesture to MouseHandler base class
Addresses code review feedback to eliminate duplication between
PointerEventHandler and TouchHandler.

- Moved _dispatchGesture method to MouseHandler base class
- Both PointerEventHandler and TouchHandler now use inherited method
- Removed NavigationCommandRevealType and MouseTargetType imports
  from pointerHandler.ts (now only needed in base class)
